PYIC-9002: Move test files to subdirectory
